js/xpconnect/tests/mochitest/test_bug393269.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=393269
-->
<head>
  <title>Test for Bug 393269</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=393269">Mozilla Bug 393269</a>
<iframe id="ifr"></iframe>
<pre id="test">
<script class="testbody" type="text/javascript">

(function () {
    /** Test for Bug 393269 **/
    var doc = $("ifr").contentDocument;
    is("UTF-8", doc.characterSet, "control, getting a property");
    doc.open();
    try {
        is("UTF-8", doc.characterSet,
           "can get a property after 1 document.open")
    } catch (e) {
        fail("Shouldn't have thrown: " + e);
        return;
    } finally {
        doc.close();
    }

    doc.open();
    try {
        is("UTF-8", doc.characterSet,
           "can get a property after 2 document.opens")
    } catch (e) {
        fail("Shouldn't have thrown: " + e);
    } finally {
        doc.close();
    }
})();
</script>
</pre>
</body>
</html>
